Consolidate spatial functions into one file
